Refactor posting action activations to provide a single entry for activation primitive or sequence action.
